gjp-assetguard-extraction-tool/requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
fpdf==1.7.2
google-cloud-storage==3.2.0
google-cloud-vision==3.10.2
openai==1.98.0
pillow==11.3.0
PyMuPDF==1.26.3
reportlab==4.1.0
requests==2.32.4
python-dotenv==1.0.0
opencv-python==4.10.0.84